Move graph to a derived store
